Move filtering query to base handler
