+++ cg/store/api/base.py
@@ -1,9 +1,13 @@
 """All models aggregated in a base class"""
-from typing import Type, List
+import datetime
+from typing import Callable, Optional, Type, List
 
 from alchy import Query, ModelBase
 from sqlalchemy import and_, func
 from dataclasses import dataclass
+from cg.store.filters.status_case_filters import CaseFilter, apply_case_filter
+from cg.store.filters.status_customer_filters import CustomerFilter, apply_customer_filter
+from cg.store.filters.status_sample_filters import SampleFilter, apply_sample_filter
 from cg.store.models import (
     Analysis,
     Application,
@@ -127,3 +131,77 @@ class BaseHandler:
                 self.Analysis.started_at == case_and_date_subquery.c.started_at,
             ),
         )
+
+    def _get_filtered_case_query(
+        self,
+        case_action: Optional[str],
+        customer_id: str,
+        data_analysis: str,
+        days: int,
+        exclude_customer_id: bool,
+        internal_id: str,
+        name: str,
+        priority: str,
+        sample_id: str,
+    ) -> Query:
+        cases_query: Query = self._get_query(table=Family)
+        filter_functions: List[Callable] = []
+
+        filter_case_order_date = None
+        if days != 0:
+            filter_case_order_date = datetime.datetime.now() - datetime.timedelta(days=days)
+            filter_functions.append(CaseFilter.GET_NEW_BY_ORDER_DATE)
+        if case_action:
+            filter_functions.append(CaseFilter.FILTER_BY_ACTION)
+        if priority:
+            filter_functions.append(CaseFilter.FILTER_BY_PRIORITY)
+        if internal_id:
+            filter_functions.append(CaseFilter.FILTER_BY_INTERNAL_ID_SEARCH)
+        if name:
+            filter_functions.append(CaseFilter.FILTER_BY_NAME_SEARCH)
+        if data_analysis:
+            filter_functions.append(CaseFilter.FILTER_BY_PIPELINE_SEARCH)
+
+        cases_query = apply_case_filter(
+            cases=cases_query,
+            filter_functions=filter_functions,
+            order_date=filter_case_order_date,
+            action=case_action,
+            priority=priority,
+            internal_id_search=internal_id,
+            name_search=name,
+            pipeline_search=data_analysis,
+        )
+
+        # customer filters
+        customer_filters: List[Callable] = []
+        if customer_id or exclude_customer_id:
+            cases_query = cases_query.join(Family.customer)
+
+        if customer_id:
+            customer_filters.append(CustomerFilter.FILTER_BY_INTERNAL_ID)
+
+        if exclude_customer_id:
+            customer_filters.append(CustomerFilter.EXCLUDE_INTERNAL_ID)
+
+        cases_query = apply_customer_filter(
+            customers=cases_query,
+            filter_functions=customer_filters,
+            customer_internal_id=customer_id,
+            exclude_customer_internal_id=exclude_customer_id,
+        )
+
+        # sample filters
+        if sample_id:
+            cases_query = cases_query.join(Family.links, FamilySample.sample)
+            cases_query = apply_sample_filter(
+                samples=cases_query,
+                filter_functions=[SampleFilter.FILTER_BY_INTERNAL_ID_PATTERN],
+                internal_id_pattern=sample_id,
+            )
+        else:
+            cases_query = cases_query.outerjoin(Family.links, FamilySample.sample)
+
+        # other joins
+        cases_query = cases_query.outerjoin(Family.analyses, Sample.invoice, Sample.flowcells)
+        return cases_query
