+++ cg/store/api/status.py
@@ -527,80 +527,6 @@ class StatusHandler(BaseHandler):
         case_data.max_tat = self._get_max_tat(links=case_obj.links)
         return case_data
 
-    def _get_filtered_case_query(
-        self,
-        case_action: Optional[str],
-        customer_id: str,
-        data_analysis: str,
-        days: int,
-        exclude_customer_id: bool,
-        internal_id: str,
-        name: str,
-        priority: str,
-        sample_id: str,
-    ) -> Query:
-        cases_query: Query = self._get_query(table=Family)
-        filter_functions: List[Callable] = []
-
-        filter_case_order_date = None
-        if days != 0:
-            filter_case_order_date = datetime.now() - timedelta(days=days)
-            filter_functions.append(CaseFilter.GET_NEW_BY_ORDER_DATE)
-        if case_action:
-            filter_functions.append(CaseFilter.FILTER_BY_ACTION)
-        if priority:
-            filter_functions.append(CaseFilter.FILTER_BY_PRIORITY)
-        if internal_id:
-            filter_functions.append(CaseFilter.FILTER_BY_INTERNAL_ID_SEARCH)
-        if name:
-            filter_functions.append(CaseFilter.FILTER_BY_NAME_SEARCH)
-        if data_analysis:
-            filter_functions.append(CaseFilter.FILTER_BY_PIPELINE_SEARCH)
-
-        cases_query = apply_case_filter(
-            cases=cases_query,
-            filter_functions=filter_functions,
-            order_date=filter_case_order_date,
-            action=case_action,
-            priority=priority,
-            internal_id_search=internal_id,
-            name_search=name,
-            pipeline_search=data_analysis,
-        )
-
-        # customer filters
-        customer_filters = []
-        if customer_id or exclude_customer_id:
-            cases_query = cases_query.join(Family.customer)
-
-        if customer_id:
-            customer_filters.append(CustomerFilter.FILTER_BY_INTERNAL_ID)
-
-        if exclude_customer_id:
-            customer_filters.append(CustomerFilter.EXCLUDE_INTERNAL_ID)
-
-        cases_query = apply_customer_filter(
-            customers=cases_query,
-            filter_functions=customer_filters,
-            customer_internal_id=customer_id,
-            exclude_customer_internal_id=exclude_customer_id,
-        )
-
-        # sample filters
-        if sample_id:
-            cases_query = cases_query.join(Family.links, FamilySample.sample)
-            cases_query = apply_sample_filter(
-                samples=cases_query,
-                filter_functions=[SampleFilter.FILTER_BY_INTERNAL_ID_PATTERN],
-                internal_id_pattern=sample_id,
-            )
-        else:
-            cases_query = cases_query.outerjoin(Family.links, FamilySample.sample)
-
-        # other joins
-        cases_query = cases_query.outerjoin(Family.analyses, Sample.invoice, Sample.flowcells)
-        return cases_query
-
     @staticmethod
     def _is_rerun(
         case_obj: Family,
